moved examples to own folder
