+++ examples/serverCH/config/server.conf
@@ -1,0 +1,48 @@
+{
+    "RootZonePublicKeyPath":        "keys/selfSignedRootDelegationAssertion.gob",
+    "ServerAddress":                {
+                                        "Type":     "TCP",
+                                        "TCPAddr":  {
+                                                        "IP":   "127.0.0.1",
+                                                        "Port": 5022,
+                                                        "Zone": ""
+                                                    }
+                                    },
+    "MaxConnections":               1000,
+    "KeepAlivePeriod":              60,
+    "TCPTimeout":                   300,
+    "TLSCertificateFile":           "config/server.crt",
+    "TLSPrivateKeyFile":            "config/server.key",
+    "MaxMsgByteLength":             65536,
+    "PrioBufferSize":               1000,
+    "NormalBufferSize":             100000,
+    "PrioWorkerCount":              2,
+    "NormalWorkerCount":            10,
+    "ActiveTokenCacheSize":         1000,
+    "ZoneKeyCacheSize":             1000,
+    "PendingSignatureCacheSize":    1000,
+    "AssertionCacheSize":           10000,
+    "PendingQueryCacheSize":        100,
+    "CapabilitiesCacheSize":        50,
+    "NotificationBufferSize":       20,
+    "NotificationWorkerCount":      2,
+    "PeerToCapCacheSize":           1000,
+    "Capabilities":                 ["urn:x-rains:tlssrv"],
+    "InfrastructureKeyCacheSize":   10,
+    "ExternalKeyCacheSize":         5,
+    "DelegationQueryValidity":      5,
+    "NegativeAssertionCacheSize":   500,
+    "AddressQueryValidity":         5,
+    "QueryValidity":                5,
+    "MaxCacheValidity":             {  
+                                        "AssertionValidity": 720,
+                                        "ShardValidity": 720,
+                                        "ZoneValidity": 720,
+                                        "AddressAssertionValidity": 720,
+                                        "AddressZoneValidity": 720
+                                    },
+    "ReapVerifyTimeout":            1800,
+    "ReapEngineTimeout":            1800,
+    "ContextAuthority":             ["."],
+    "ZoneAuthority":                ["ch"]
+}
