+++ examples/serverCH/keys/zonePrivate.key
@@ -1,0 +1,1 @@
+80e1a328b908c2d6c2f10659355b15618ead2e42acf1dfcf39488fc7006c444e2245137bcb058f799843bb8c6df31927b547e4951142b99ae97c668b076e9d84
