+++ examples/serverCH/keys/zonePublic.key
@@ -1,0 +1,1 @@
+2245137bcb058f799843bb8c6df31927b547e4951142b99ae97c668b076e9d84
