--- serverCHPub/zoneFiles/chZoneFile.txt
@@ -1,6 +1,0 @@
-:Z: ch . [
-                :S: [
-                        :A: ch  [ :ip4: 178.209.53.76 ]
-                        :A: ethz [ :ip4: 129.132.128.139 ]
-                    ]
-            ]
