--- serverRoot/config/server.key
@@ -1,27 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEowIBAAKCAQEA1oisZHKMAqyPsxrXHBdPu+UHUGy2vrQZFCv9AoKvcRp2GBd9
-fi7hKUKeizBVXxfy328iTyXGjN87+Nisl38nZgb5xEeWbl1VHa3dwB74CGE7IyL2
-2kFwZj26LN4Pmw/IaqHdkXheOoxDI3k3w4A000ISzzvNh/BUPOJ9Ha1ill1QwqMj
-QbtanJ1KbLAqNmrXVUTyuB3q84aKYs4tlrFLfID43Z3wv1gCiEEAquEJRoOeq3li
-PT3Z7dA/mEG7KNH41J28JVf98AxUyRRIP5CktHxFgQrzKQcJnNYlo07q8fjUoDJm
-ITfgfiuDTluigAOcXQGBsNjYKVVX+LBizQqAiQIDAQABAoIBABCny7qAqh4LB50A
-GHM13BraTRcRWZ60nL3ULHBldPly9HgqWrPRL+nkIirsNbSMR27RUfiTFjcZna5U
-JBIe7RAeuoViiWtIE7OaK/0eamM8KTsxG7kWG4PUhv+0vcFsGEMvDeSdU7N38D5w
-ysOC0JcWKh+spOhI4EFZ1BzZuS2zLIevFqwd4xScW5pbW3cTiuNWEhqPfyu9C06w
-g9XEEtmwZWdjLGK0zp8jVdcDQgW5GvlwpXJJGmzvB2DHZTOTPqiUnSg8rnYc9o6M
-gETY0YAFn0QYUbPpWX21n9YRNks9xLFRum91JlcFlPQbA7+yv6eLqYC3/Hnbcgfk
-E6Le4AECgYEA83Z/S8BGUc0Np93TrhtH+TtJbSHTzzZxtTfZPGY0lhxBjrW/1Mk3
-ubIlC9prlDSs3iD19melQC4gLkJqCowihLQgBpyeEElf7KvI65/J5mI1+nUwRhio
-9mm6fhN8Y0pV6ekvlBhHetk9abisyoA5MgE9j70NHt+3SHUBTkE7eqECgYEA4ZTQ
-M4VAv21gIJwyVIl5tk8Q++UaHzLOcMDnISzefSLm6CdvoU0Cbnlun8Ps89ASShLI
-tb/u54h4TbjsUd8T+Fy0KZ33lCbnV29A9LUOEmem7sinqtandqsaKv+rRDXsLZQS
-VGPtsLRuTWygw1WaJWmvxD9k0iaaPgnRHdthZOkCgYEA8X2CerL3jUAl4XJo+3PY
-CmFCTduV6x2H4EbLVdvTIWpZpKKm6vvEKI+Jq541w0YUiiiLkKPyCB04kmoSIPAr
-dRHuCmDjcCg329vxabaPX9Pe0tXW0lC3l+P7tf7CDQK5AVHpfa4viK1e+zTRP/O7
-x9vvIlpVX2BsvpRAfUy7IMECgYA21dDipUK/Bdx3MOD/Mcf3bDLpuKBn8xbBy/Kj
-Wt/etWxCBMwa/rhXmIcuqRbhkbPQc858d+KgP8lWw1wW7s5Q8mANOcwkJGzaET5N
-/fEkwjOPGhG/8pwUubULvMiqNR1rWvI1t3DhIospoP1XXmixLYa/gtWETEP0TriZ
-tGlxeQKBgEJ9m4lyJMbnZr6+dMhd1DTPwXspHbb8HMgSOwoVc5OVPFK4lEHWgu1Y
-cXT6frGOT8Ndp2ehiZehQSkwQSfYpSWZ9iyzcTxXPj2/tTF37rUSZv0ZmxM3Vapr
-G89AZsRJXcr43Nn9JdL060Scw3n/HiTiGT0Tm1nu6ntkXUlkpFt5
------END RSA PRIVATE KEY-----
