--- serverRootPub/keys/_rootPrivate.key
@@ -1,1 +1,0 @@
-2976132013b35e912a2a1072d6b198f46d45feebc3a382dd8b0511ee111802532e4dc53dd4f04dad288d51d1de11b9cca319b3f4dae6d191313727df9d9c1225
