--- serverRootPub/keys/rootPrivate.key
@@ -1,1 +1,0 @@
-d1841498580dd5512201275f842ec2e6df6a946f59c5e82aac6be6b2302ba3705f5282f6866997dca102938e4365f701f8c48cc074035df8c557b7e674a31ffa
